+++ server/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Breaking
+
+- Formulas now accepts `null`. Old formulas referencing a column named `null` must do so by using `[null]` instead.
+
 ### Fixed
 
 - Mongo: step `todate` properly supports two-digit years (`%y`).